# backend/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ShortlistSummaryDTO": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"job": {
						"type": "string"
					},
					"jobTitle": {
						"type": "string"
					},
					"runName": {
						"type": "string"
					},
					"model": {
						"type": "string"
					},
					"totalApplicants": {
						"type": "number",
						"format": "double"
					},
					"shortlistCount": {
						"type": "number",
						"format": "double"
					},
					"topMatchScore": {
						"type": "number",
						"format": "double"
					},
					"topCandidateName": {
						"type": "string"
					},
					"screeningStartedAt": {
						"type": "string"
					},
					"screeningCompletedAt": {
						"type": "string"
					},
					"screeningDurationSeconds": {
						"type": "number",
						"format": "double"
					},
					"createdBy": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"job",
					"jobTitle",
					"runName",
					"model",
					"totalApplicants",
					"shortlistCount",
					"topMatchScore",
					"topCandidateName",
					"createdBy",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ShortlistListResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ShortlistSummaryDTO"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"total",
					"page",
					"pageSize",
					"totalPages",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ShortlistSelectItemDTO": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"runName": {
						"type": "string"
					},
					"job": {
						"type": "string"
					},
					"jobTitle": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"runName",
					"job",
					"jobTitle",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ShortlistSelectResponse": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/ShortlistSelectItemDTO"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ShortlistWeightCriterionDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"value": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"label",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ShortlistCriterionAssessmentDTO": {
				"properties": {
					"label": {
						"type": "string"
					},
					"weightPct": {
						"type": "number",
						"format": "double"
					},
					"score": {
						"type": "number",
						"format": "double"
					},
					"weightedScore": {
						"type": "number",
						"format": "double"
					},
					"summary": {
						"type": "string"
					},
					"evidence": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"label",
					"weightPct",
					"score",
					"weightedScore"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ShortlistRecommendation": {
				"type": "string",
				"enum": [
					"Strong Reject",
					"Reject",
					"Consider",
					"Shortlist",
					"Strong Shortlist"
				]
			},
			"ShortlistResultEntryDTO": {
				"properties": {
					"candidateRank": {
						"type": "number",
						"format": "double"
					},
					"applicantEmail": {
						"type": "string"
					},
					"fullName": {
						"type": "string"
					},
					"matchScore": {
						"type": "number",
						"format": "double"
					},
					"confidenceScore": {
						"type": "number",
						"format": "double"
					},
					"skillsScore": {
						"type": "number",
						"format": "double"
					},
					"experienceScore": {
						"type": "number",
						"format": "double"
					},
					"educationScore": {
						"type": "number",
						"format": "double"
					},
					"relevanceScore": {
						"type": "number",
						"format": "double"
					},
					"criterionAssessments": {
						"items": {
							"$ref": "#/components/schemas/ShortlistCriterionAssessmentDTO"
						},
						"type": "array"
					},
					"criticalRequirementGap": {
						"type": "boolean"
					},
					"strengths": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"gapsOrRisks": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"finalRecommendation": {
						"$ref": "#/components/schemas/ShortlistRecommendation"
					},
					"summaryExplanation": {
						"type": "string"
					}
				},
				"required": [
					"candidateRank",
					"applicantEmail",
					"fullName",
					"matchScore",
					"confidenceScore",
					"skillsScore",
					"experienceScore",
					"educationScore",
					"relevanceScore",
					"criticalRequirementGap",
					"strengths",
					"gapsOrRisks",
					"finalRecommendation",
					"summaryExplanation"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ShortlistDTO": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"job": {
						"type": "string"
					},
					"jobTitle": {
						"type": "string"
					},
					"runName": {
						"type": "string"
					},
					"model": {
						"type": "string"
					},
					"totalApplicants": {
						"type": "number",
						"format": "double"
					},
					"shortlistCount": {
						"type": "number",
						"format": "double"
					},
					"weightCriteria": {
						"items": {
							"$ref": "#/components/schemas/ShortlistWeightCriterionDTO"
						},
						"type": "array"
					},
					"screeningResults": {
						"items": {
							"$ref": "#/components/schemas/ShortlistResultEntryDTO"
						},
						"type": "array"
					},
					"shortlist": {
						"items": {
							"$ref": "#/components/schemas/ShortlistEntryDTO"
						},
						"type": "array"
					},
					"instructions": {
						"type": "string"
					},
					"screeningStartedAt": {
						"type": "string"
					},
					"screeningCompletedAt": {
						"type": "string"
					},
					"screeningDurationSeconds": {
						"type": "number",
						"format": "double"
					},
					"createdBy": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"_id",
					"job",
					"jobTitle",
					"runName",
					"model",
					"totalApplicants",
					"shortlistCount",
					"weightCriteria",
					"screeningResults",
					"shortlist",
					"createdBy",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ShortlistResponse": {
				"properties": {
					"data": {
						"$ref": "#/components/schemas/ShortlistDTO"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"data",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateShortlistRequest": {
				"properties": {
					"jobId": {
						"type": "string"
					},
					"runName": {
						"type": "string"
					},
					"jobTitle": {
						"type": "string"
					},
					"model": {
						"type": "string"
					},
					"totalApplicants": {
						"type": "number",
						"format": "double"
					},
					"shortlistCount": {
						"type": "number",
						"format": "double"
					},
					"weightCriteria": {
						"items": {
							"$ref": "#/components/schemas/ShortlistWeightCriterionDTO"
						},
						"type": "array"
					},
					"screeningResults": {
						"items": {
							"$ref": "#/components/schemas/ShortlistResultEntryDTO"
						},
						"type": "array"
					},
					"shortlist": {
						"items": {
							"$ref": "#/components/schemas/ShortlistEntryDTO"
						},
						"type": "array"
					},
					"instructions": {
						"type": "string"
					},
					"screeningStartedAt": {
						"type": "string"
					},
					"screeningCompletedAt": {
						"type": "string"
					},
					"screeningDurationSeconds": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"jobId",
					"jobTitle",
					"totalApplicants",
					"shortlistCount",
					"screeningResults",
					"shortlist"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateShortlistCandidateStatusRequest": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"shortlisted",
							"rejected",
							"interview",
							"exam",
							"assessment",
							"practical"
						]
					}
				},
				"required": [
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteShortlistResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "backend-api",
		"version": "1.0.0",
		"description": "Simple backend API with Hello World endpoint using TypeScript, tsoa, and Mongoose",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Bikorimana Saveur"
		}
	},
	"paths": {
...
		"/shortlists": {
			"get": {
				"operationId": "ListShortlists",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShortlistListResponse"
								}
							}
						}
					}
				},
				"description": "List shortlists with pagination, optionally filtered by job.",
				"tags": [
					"Shortlists"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "jobId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateShortlist",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShortlistResponse"
								}
							}
						}
					}
				},
				"description": "Persist a screening run result as a shortlist record.",
				"tags": [
					"Shortlists"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateShortlistRequest"
							}
						}
					}
				}
			}
		},
		"/shortlists/select": {
			"get": {
				"operationId": "ListShortlistSelect",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShortlistSelectResponse"
								}
							}
						}
					}
				},
				"description": "Lightweight shortlist rows for selectors.",
				"tags": [
					"Shortlists"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "jobId",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 100,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/shortlists/{id}": {
			"get": {
				"operationId": "GetShortlist",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShortlistResponse"
								}
							}
						}
					}
				},
				"description": "Get a single shortlist by ID (includes full screening results).",
				"tags": [
					"Shortlists"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteShortlist",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/DeleteShortlistResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Shortlists"
				],
				"security": [
					{
						"jwt": [
							"recruiter",
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/shortlists/{id}/candidates/{email}": {
			"patch": {
				"operationId": "UpdateShortlistCandidateStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ShortlistResponse"
								}
							}
						}
					}
				},
				"description": "Delete a shortlist permanently.",
				"tags": [
					"Shortlists"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateShortlistCandidateStatusRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}